refactor: Rename components. Style tweaks.
